Changed the -hash method so that it conforms to NSObject protocol
git-svn-id: http://basetenframework.org/svn/trunk@1390 fe1aa653-7022-0410-af67-bafd91c7e665

+++ Sources/BXAbstractDescription.h
@@ -36,7 +36,7 @@
 }
 
 - (NSString *) name;
-- (unsigned int) hash;
+- (NSUInteger) hash;
 - (id) initWithName: (NSString *) aName;
 //- (void) setName: (NSString *) aName;
 @end
